requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
pandas==2.2.3
openpyxl==3.1.2
gspread==5.12.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
Werkzeug==3.0.1
gunicorn==22.0.0
psutil==5.9.6
